+++ src/runtime/rhi/stream_tag.h
@@ -4,7 +4,7 @@
 
 namespace luisa::compute {
 
-enum class StreamTag : uint8_t {
+enum class StreamTag : uint32_t {
     GRAPHICS,// capable of graphics, compute, and copy commands
     COMPUTE, // capable of compute and copy commands
     COPY     // only copy commands
